feat: CDSH-149: Add backward compatibility with 1.27.x.
