+++ README.MD
@@ -3,19 +3,19 @@ RunMultiplePipelines plugin adds a new custom stage to Spinnaker to trigger mult
 ### Version Compatibility
 Minimum tested version is 1.26.x
 
-| Spinnaker Version  | Plugin Version |
-|:------------- | :--------- |
-| <= 1.27.x | 1.0.2 |
-| <= 1.27.x | 1.0.3 |
-| <= 1.27.x | 1.0.4 |
-| <= 1.27.x | 1.0.5 |
-| <= 1.27.x | 1.0.6 |
-| <= 1.27.x | 1.0.7 |
-| <= 1.27.x | 1.0.8 |
-| <= 1.27.x | 1.0.9 |
-| >= 1.27.x | 1.1.0 |
-| >= 1.30.x | 1.1.1 |
-
+| Spinnaker Version | Plugin Version |
+|:------------------|:---------------|
+| <= 1.27.x         | 1.0.2          |
+| <= 1.27.x         | 1.0.3          |
+| <= 1.27.x         | 1.0.4          |
+| <= 1.27.x         | 1.0.5          |
+| <= 1.27.x         | 1.0.6          |
+| <= 1.27.x         | 1.0.7          |
+| <= 1.27.x         | 1.0.8          |
+| <= 1.27.x         | 1.0.9          |
+| >= 1.27.x         | 1.1.0          |
+| >= 1.30.x         | 1.1.1          |
+| >= 1.27.x         | 1.1.2          |
 
 NOTE: The plugin is not actively tested in all compatible versions with all variants, but is expected to work in the above.
 
@@ -37,7 +37,7 @@ spinnakerConfig:
             plugins:
               Armory.RunMultiplePipelines:
                 enabled: true
-                version: 1.1.1
+                version: 1.1.2
           repositories:
             runMultiplePipelinesRepo:
               url: https://raw.githubusercontent.com/armory/multiple-pipelines-plugin-releases/main/plugins.json
@@ -47,7 +47,7 @@ spinnakerConfig:
           plugins:
             Armory.RunMultiplePipelines:
             enabled: true
-            version: 1.1.1
+            version: 1.1.2
             extensions:
               armory.runMultiplePipelinesStage:
                 enabled: true
