+++ gradle.properties
@@ -1,4 +1,4 @@
 spinnakerGradleVersion=8.16.0
-orcaVersion=8.32.0
+orcaVersion=8.25.0
 
 artifactory_token=
