+++ orca-run-multiple-pipelines/src/main/java/io/armory/plugin/smp/tasks/ParsePipelinesYamlTask.java
@@ -3,6 +3,7 @@ package io.armory.plugin.smp.tasks;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.graph.MutableGraph;
+import com.netflix.spinnaker.kork.plugins.api.spring.ExposeToApp;
 import com.netflix.spinnaker.orca.api.pipeline.Task;
 import com.netflix.spinnaker.orca.api.pipeline.TaskResult;
 import com.netflix.spinnaker.orca.api.pipeline.models.ExecutionStatus;
@@ -21,6 +22,7 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
+@ExposeToApp
 @Component
 public class ParsePipelinesYamlTask implements Task {
